Refactored CLI functionality and error handling in forc-debug crate.
