+++ forc-plugins/forc-debug/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod cli;
 pub mod error;
 pub mod names;
 pub mod server;
